Fix bug gaia_id in nasa and epic
Starting from October 2025 the column gaia_id
in the nasa and epic catalogs has been splitted in two:
- gaia_dr2_id
- gaia_dr3_id

Both for nasa and epic, it adds the rename of the original
column from gaia_dr3_id (in nasa and epic) to gaia_id (in emc).
This should be also backward compatible.